Remove redux-related file for BlockingModal
